Remove all spl-git packages

+++ build.sh
@@ -138,14 +138,18 @@ generate_package_files() {
     fi
 
     # Finally, generate the update packages ...
-    if [[ "${kernel_name}" == "common" ]] || [[ "${kernel_name}" == "common-git" ]]; then
+    
+    # skip spl for common-git
+    if [[ "${kernel_name}" == "common" ]]; then
         msg2 "Removing old spl-utils patches (if any)"
         run_cmd_no_output "rm -f ${spl_utils_pkgbuild_path}/*.patch"
         msg2 "Copying spl-utils patches (if any)"
         run_cmd_no_output "find ${script_dir}/src/spl-utils -name \*.patch -exec cp {} ${spl_utils_pkgbuild_path} \;"
         msg2 "Creating spl-utils PKGBUILD"
         run_cmd_no_output "source ${script_dir}/src/spl-utils/PKGBUILD.sh"
+    fi
 
+    if [[ "${kernel_name}" == "common" ]] || [[ "${kernel_name}" == "common-git" ]]; then
         msg2 "Removing old zfs-utils patches (if any)"
         run_cmd_no_output "rm -f ${zfs_utils_pkgbuild_path}/*.patch"
         msg2 "Copying zfs-utils patches (if any)"
@@ -161,17 +165,19 @@ generate_package_files() {
         msg2 "Copying zfs-utils.initcpio.install"
         run_cmd_no_output "cp ${script_dir}/src/zfs-utils/zfs-utils.initcpio.install ${zfs_utils_pkgbuild_path}/zfs-utils.initcpio.install"
     elif [[ "${kernel_name}" == "dkms" ]]; then
-        msg2 "Removing old spl patches (if any)"
-        run_cmd_no_output "rm -f ${spl_dkms_pkgbuild_path}/*.patch"
-        msg2 "Copying spl patches (if any)"
-        run_cmd_no_output "find ${script_dir}/src/spl-dkms -name \*.patch -exec cp {} ${spl_dkms_pkgbuild_path} \;"
-        msg2 "Copying 60-spl-dkms-install.hook"
-        run_cmd_no_output "cp ${script_dir}/src/spl-dkms/60-spl-dkms-install.hook ${spl_dkms_pkgbuild_path}/"
-        msg2 "Copying spl-dkms-alpm-hook"
-        run_cmd_no_output "cp ${script_dir}/src/spl-dkms/spl-dkms-alpm-hook ${spl_dkms_pkgbuild_path}/"
-        msg2 "Creating spl-dkms PKGBUILD"
-        run_cmd_no_output "source ${script_dir}/src/spl-dkms/PKGBUILD.sh"
-
+        # skip spl for git packages
+        if [[ ! ${archzfs_package_group} =~ -git$ ]]; then
+            msg2 "Removing old spl patches (if any)"
+            run_cmd_no_output "rm -f ${spl_dkms_pkgbuild_path}/*.patch"
+            msg2 "Copying spl patches (if any)"
+            run_cmd_no_output "find ${script_dir}/src/spl-dkms -name \*.patch -exec cp {} ${spl_dkms_pkgbuild_path} \;"
+            msg2 "Copying 60-spl-dkms-install.hook"
+            run_cmd_no_output "cp ${script_dir}/src/spl-dkms/60-spl-dkms-install.hook ${spl_dkms_pkgbuild_path}/"
+            msg2 "Copying spl-dkms-alpm-hook"
+            run_cmd_no_output "cp ${script_dir}/src/spl-dkms/spl-dkms-alpm-hook ${spl_dkms_pkgbuild_path}/"
+            msg2 "Creating spl-dkms PKGBUILD"
+            run_cmd_no_output "source ${script_dir}/src/spl-dkms/PKGBUILD.sh"
+        fi
         msg2 "Removing old zfs patches (if any)"
         run_cmd_no_output "rm -f ${zfs_dkms_pkgbuild_path}/*.patch"
         msg2 "Copying zfs patches (if any)"
@@ -183,14 +189,17 @@ generate_package_files() {
         zfs_headers_conflicts=${zfs_headers_conflicts_all/"'${zfs_pkgname}-headers'"}
         spl_headers_conflicts=${spl_headers_conflicts_all/"'${spl_pkgname}-headers'"}
 
-        msg2 "Removing old spl patches (if any)"
-        run_cmd_no_output "rm -f ${spl_pkgbuild_path}/*.patch"
-        msg2 "Copying spl patches (if any)"
-        run_cmd_no_output "find ${script_dir}/src/spl -name \*.patch -exec cp {} ${spl_pkgbuild_path} \;"
-        msg2 "Creating spl PKGBUILD"
-        run_cmd_no_output "source ${script_dir}/src/spl/PKGBUILD.sh"
-        msg2 "Creating spl.install"
-        run_cmd_no_output "source ${script_dir}/src/spl/spl.install.sh"
+        # skip spl for git packages
+        if [[ ! ${archzfs_package_group} =~ -git$ ]]; then
+            msg2 "Removing old spl patches (if any)"
+            run_cmd_no_output "rm -f ${spl_pkgbuild_path}/*.patch"
+            msg2 "Copying spl patches (if any)"
+            run_cmd_no_output "find ${script_dir}/src/spl -name \*.patch -exec cp {} ${spl_pkgbuild_path} \;"
+            msg2 "Creating spl PKGBUILD"
+            run_cmd_no_output "source ${script_dir}/src/spl/PKGBUILD.sh"
+            msg2 "Creating spl.install"
+            run_cmd_no_output "source ${script_dir}/src/spl/spl.install.sh"
+        fi
 
         msg2 "Removing old zfs patches (if any)"
         run_cmd_no_output "rm -f ${zfs_pkgbuild_path}/*.patch"
